Move location of process-fews-event-code-test-utils
