+++ testing/function-tests/ProcessFewsEventCode/test.timeseriesIgnoredWorkflow.js
@@ -3,7 +3,7 @@ module.exports = describe('Tests for import timeseries ignored workflows', () =>
   const Context = require('../mocks/defaultContext')
   const ConnectionPool = require('../../../Shared/connection-pool')
   const CommonTimeseriesTestUtils = require('../shared/common-timeseries-test-utils')
-  const ProcessFewsEventCodeTestUtils = require('../shared/process-fews-event-code-test-utils')
+  const ProcessFewsEventCodeTestUtils = require('./process-fews-event-code-test-utils')
   const sql = require('mssql')
 
   let context
